fix(tools): validate deploy runId as integer and warn on secret-bearing deploy log
Copilot review on PR #258 found two issues, both verified against the code:

- `runId: z.number()` on get_stack_deploy/delete_stack_deploy/get_stack_deploy_log
  accepted a non-integer like 12.5, which Dockhand's upstream handler silently
  truncates via parseInt(). For delete_stack_deploy this could delete the wrong
  run's deploy record. Fixed with `.int()` on all three (list_stack_deploys has
  no runId and is untouched).
- get_stack_deploy_log can return secrets that survived Dockhand's own log
  redaction (see the tool's inline comment), but its derived description read
  as a plain "fetch the log" with no hint of that. Added an operator-safety
  suffix (description-suffixes.ts category 2), matching the existing
  get_stack_env_raw pattern.

Tests exercise the actual Zod schema via z.object(schema).safeParse() rather
than a source-text regex, since a regex match on `.int()` only proves the
token is present, not that Zod rejects a non-integer at runtime. Verified the
counter-check by hand: reverting `.int()` makes exactly the 3 "rejects 12.5"
schema tests and the 3 regex assertions fail, for the expected reason.

+++ src/tools/stacks.ts
@@ -885,7 +885,7 @@ export function registerStackTools(server: McpServer, client: DockhandClient): v
   registerTool(server, 'get_stack_deploy',
     {
       name: z.string().describe('Stack name'),
-      runId: z.number().describe('Deploy run id (from list_stack_deploys)'),
+      runId: z.number().int().describe('Deploy run id (from list_stack_deploys)'),
     },
     async ({ name, runId }) => {
       return jsonResponse(await client.get(`/api/stacks/${encodePath(name)}/deploys/${encodePath(runId)}`));
@@ -895,7 +895,7 @@ export function registerStackTools(server: McpServer, client: DockhandClient): v
   registerTool(server, 'delete_stack_deploy',
     {
       name: z.string().describe('Stack name'),
-      runId: z.number().describe('Deploy run id (from list_stack_deploys). Fails with a 409 if the run has not finished yet (status is still queued/running) — retry once it completes.'),
+      runId: z.number().int().describe('Deploy run id (from list_stack_deploys). Fails with a 409 if the run has not finished yet (status is still queued/running) — retry once it completes.'),
     },
     async ({ name, runId }) => {
       return jsonResponse(await client.delete(`/api/stacks/${encodePath(name)}/deploys/${encodePath(runId)}`));
@@ -905,7 +905,7 @@ export function registerStackTools(server: McpServer, client: DockhandClient): v
   registerTool(server, 'get_stack_deploy_log',
     {
       name: z.string().describe('Stack name'),
-      runId: z.number().describe('Deploy run id (from list_stack_deploys)'),
+      runId: z.number().int().describe('Deploy run id (from list_stack_deploys)'),
     },
     async ({ name, runId }) => {
       // This is the most sensitive of the four deploy-history endpoints — it can
